requirements.txt
Pillow==2.1.0
spidev==3.4
pyppeteer==0.0.23
aiocron==1.3
websockets==6.0
RPi.GPIO==0.7.0
numpy==1.17.4
